==> Yukon-KuskokwimChinook/precipitation/0_GRFR.Rmd <==
---
title: "0_GRFR"
output: 
  html_document:
    code_folding: TRUE
date: "2022-08-30"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(sf)
library(lubridate)
library(tidyverse)
library(googlesheets4)
library(corrplot)
library(gridExtra)
library(GGally)

```

Extract modeled streamflow for watershed outlets for analysis sites. These include watersheds with productivity estimates and major spawning populations for the genetic groupings. Watersheds were extracted for all major spawning populations in the Yukon so can extract modeled streamflow for all of these sites. This will be useful for plotting variability in maximum fall streamflow at finer versus coarser spatial scales (e.g. major spawning areas in the lower Yukon mgmt area).

1. Load point file that has outlets for different watersheds that have been referenced to the MERIT basins stream network. These have been checked against the V01 of MERIT hydro streams and catchments since that version is linked to the GRFR values (per emails from Josh).
2. Intersect with catchments to get COMIDs. Use V01 catchments.
3. Read in GRFR file created by Josh and extract the modeled flow for all 40 years for each watershed outlet.
4. Create annual metrics of maximum fall streamflow for all 40 years and all sites. 
5. For genetic groups, compare peak streamflow time series for major spawning areas to see if one is representative or if some sort of average should be used. 

# Get Maximum Daily Fall Streamflow for watershed outlets

```{r load spatial data}

pts <- read_sf(dsn = "W:\\GIS\\AYK_Chinook\\AYK_Chinook.gdb", layer = "chinook_sites_fromR_shiftedToMB")

#filter to only include the escapement monitoring sites used in the analysis 
# and the major spawning areas in the Yukon watershed.
pts <- pts %>% 
  filter(analysis_site == 1 | analysis_site == 2 & Management_Area == "Kuskokwim") %>% 
  arrange(Management_Area, Location)

#removed duplicates in the file that were both escapement locations and major spawning areas in Brown et al.
```

```{r eval = FALSE}
cats <- read_sf(dsn = "W:\\GIS\\MERIT-Basins\\V01", layer = "cat_pfaf_8_MERIT_Hydro_v07_Basins_v01")

#filter to AK only (currently pfaf 8, which is entire arctic)
cats <- cats %>% 
  filter(COMID < 82000000)

st_is_valid(cats)
cats2 <- st_make_valid(cats)

```

Intersect the points with the catchments 

```{r get comids}
pts2 <- st_join(pts, cats2)
pts2 %>% distinct(COMID) #49 just like points so no spatial dups

```

Read in files provided by Josh for different hydrobasins. Josh provided a few python scripts to create this dataset: extracting the daily average discharge from the original netcdf files and converting to csv, combining the modeled flow data from stream reaches within grouped hydrobasins (I supplied these to Josh), and fixing the data column in the combined table. 

Pretty amazing that R can read in a 3.6 GB file and filter it pretty efficiently!


```{r grfr file}
grfr <- read_csv("W:\\GIS\\GRFR\\GRFR_Qout_by_HydroBasin_7-28-2022\\GRFR_Qout_by_HydroBasin_7-28-2022.csv")

#filter to the comids that match the watershed outlet points
comid_filter <- pts2 %>% 
  pull(COMID) %>% 
  as.character()

grfr2 <- grfr %>%
  select(DATE, any_of(comid_filter)) %>% 
  pivot_longer(cols = -DATE, names_to = "COMID", values_to = "daily_q") %>% 
  mutate(COMID = as.numeric(COMID))

grfr2 %>% distinct(COMID)
grfr2 %>% distinct(year(DATE))
  
#save file for 1_streamflow_validation script so data can be compared to measured streamflow.
# add attribute information for sites first.
grfr2 <- left_join(grfr2, pts2 %>% st_drop_geometry() %>% select(COMID, Management_Area, Location, location_short)) 
saveRDS(grfr2, file = "data/grfr_40yrs_46sites.rds")
```

Create new annual metrics for each of these sites. There are a few that can be validated against actual streamflow data. The rest will be used as one set of covariates for the analysis.

```{r summarize annual flow metrics}

grfr2 <- readRDS("data/grfr_40yrs_46sites.rds")

grfr_maxfallq <- grfr2 %>% 
  filter(month(DATE) %in% 8:11) %>% 
  group_by(Management_Area, Location, location_short, Year = year(DATE)) %>% 
  summarize(max_fall_q = max(daily_q))

```


# Compare Data to Analysis Sites and Select appropriate spawning areas for genetic substocks

Check concordance among major spawning areas in genetic subgroups of Yukon. Read in google drive file that Megan started with site names and where I added the major spawning areas for the Lower and Middle Yukon and also the Canadian Yukon genetic substocks. 

The sites data frame should be expanded so we have all site names for analysis (Population), but also all of the major spawning areas that go to the genetic substocks. There are 4 sites that are included as separate populations (Chena, Salcha, Gisasa, and EF Andreafsky), but are also included as major spawning areas in the Lower and Middle mainstem. 

```{r read google sheet of sites}
sites <- read_sheet("https://docs.google.com/spreadsheets/d/1uS7iIqob7lP5zv94FCtnjfDlQgjBr-D1LCq4RGuLcbI/edit#gid=0",
                    na = "NA")

sites <- sites %>% 
  separate_rows(Major_Spawning_Areas, sep = ",") %>% 
  mutate(Major_Spawning_Areas = gsub(" ", "", Major_Spawning_Areas))

sites %>% distinct(Major_Spawning_Areas) %>% arrange(Major_Spawning_Areas)
sites %>% filter(Region == "Yukon (US)") %>% arrange(Major_Spawning_Areas)

grfr3 <- left_join(sites, grfr_maxfallq, by = c("Major_Spawning_Areas" = "location_short")) 

genetic_groups <- sites %>% filter(Region == "Yukon (CA)" | Region == "Yukon (US)" & grepl("Mainstem", Population)) %>% 
  distinct(Population, Region)

```

Explore dynamics of fall peak flow for major spawning areas within Canadian genetic substocks. For Lower mainstem, Stewart, Upper lakes and mainstem, and White-Donjek, there is only one major spawning area within the larger substock. Evaluate those with more than one major spawning area. Note that the two sites in the middle mainstem are dramatically different in size: mainstem Yukon and Tatchun Creek. 

```{r time series of max fall q by canadian spawning areas}
grfr3 %>%  
  filter(Population %in% c("Carmacks", "MidMain", "Pelly", "Teslin")) %>% 
  ggplot(aes(x = Year, y = max_fall_q, color = Major_Spawning_Areas)) +
  geom_line() +
  facet_wrap(~Population, scales = "free_y")
```


```{r pairplots of spawning areas within populations}
list1 <- grfr3 %>%
  filter(Population %in% c("Carmacks", "MidMain", "Pelly", "Teslin")) %>%
  group_by(Population) %>%
  group_split() %>%
  lapply(., function(x) x %>%
           select(Population, Year, Major_Spawning_Areas, max_fall_q) %>%
           pivot_wider(names_from = Major_Spawning_Areas, values_from = max_fall_q) %>%
           select(where(~!all(is.na(.x)))))

#not as good, just raw correlations
lapply(list1, function(x) x %>%
                  select(-Population, -Year) %>%
                  cor(.) %>%
                  corrplot.mixed(.))

#data + correlations
pdf("output/grfrFallQ_pairs_CanYukon.pdf")
lapply(list1, function(x) {
  title = unique(x$Population)
  x %>% 
    select(-Population, -Year) %>% 
    ggpairs(title = title) }) 
dev.off()

```



```{r correlation matrix of Canadian Yukon spawning areas}
groups = grfr3 %>% 
  filter(Population %in% c("Carmacks", "MidMain", "Pelly", "Teslin")) %>% 
  distinct(Major_Spawning_Areas) %>% 
  pull(Major_Spawning_Areas)

cor_mat <- sapply(1:length(groups), function(i)
    sapply(1:length(groups), function(j)
        cor(x = grfr3$max_fall_q[grfr3$Major_Spawning_Areas == groups[i]], 
            y = grfr3$max_fall_q[grfr3$Major_Spawning_Areas == groups[j]])))

rownames(cor_mat) <- groups
colnames(cor_mat) <- groups

cor_mat
```

We also wanted to examine the concordance/heterogeneity of fall streamflow when using major genetic subgroups in the US mainstem Yukon.

```{r time series of max fall q in lower and middle yukon}

grfr3 %>%  
  filter(Population %in% c("MainstemLower", "MainstemMid")) %>% 
  ggplot(aes(x = Year, y = max_fall_q, color = Major_Spawning_Areas)) +
  geom_line() +
  # scale_y_log10() +
  facet_wrap(~Population, scales = "free_y", ncol = 1)

#which years had the highest fall peakflow in the 40 year dataset?
grfr3 %>% 
  group_by(Major_Spawning_Areas) %>% 
  mutate(maxQ = max(max_fall_q)) %>% 
  filter(max_fall_q == maxQ) %>% 
  ungroup() %>% 
  count(Year) %>% 
  arrange(desc(n))
```

```{r pairplots of max fall q for lower and middle yukon}
list2 <- grfr3 %>% 
  filter(Population %in% c("MainstemLower", "MainstemMid")) %>% 
  group_by(Population) %>% 
  group_split() %>% 
  lapply(., function(x) x %>% 
           select(Population, Year, Major_Spawning_Areas, max_fall_q) %>% 
           pivot_wider(names_from = Major_Spawning_Areas, values_from = max_fall_q) %>% 
           select(where(~!all(is.na(.x)))))

#data + correlations
pdf("output/grfrFallQ_pairs_Yukon.pdf")
lapply(list2, function(x) {
  title = unique(x$Population)
  x %>% 
    select(-Population, -Year) %>% 
    ggpairs(title = title) }) 
dev.off()

#most of these correlations are driven by large outliers, try removing 2019
lapply(list2, function(x) {
  title = unique(x$Population)
  x %>% 
    filter(!Year == 2019) %>% 
    select(-Population, -Year) %>% 
    ggpairs(title = title) }) 

```



# Final dataset of modeled streamflow for all analysis sites

Share above data with group and decide on a major spawning area to use for each genetic substock in Canada.

(See notes in Glofas script, recommend using Teslin, mainstem Yukon, Ross, and Big Salmon based on their importance as spawning areas within the genetic groups that have more than one spawning major spawning area.)

```{r}

saveRDS(grfr3, "data/final_data/grfr_streamflow.rds")

grfr3 %>% distinct(Population)
grfr3 %>% distinct(Major_Spawning_Areas)

```
